CA-174892: scsiutil.getSCSIid() fails if multipath device is given
From: Kostas Ladopoulos <Konstantinos.Ladopoulos@citrix.com>

When getSCSIid() is passed a multipath device (e.g. /dev/dm-2) it fails,
as rawdev() strips any numbers in the end of the device name. Firstly,
rawdev() is modified to return device names that contain 'dm-' untouched.
Secondly, getSCSIid() uses 'path' with the new version of the command,
and only searches for the raw device name in the fallback call.

Signed-off-by: Kostas Ladopoulos <Konstantinos.Ladopoulos@citrix.com>